creating classes to suit function
